Replace pv_checksum_partial's return type

+++ include/pv/checksum.h
@@ -7,7 +7,7 @@ extern "C" {
 #include <stdint.h>
 
 uint16_t pv_checksum(const void* start, uint32_t size);
-uint16_t pv_checksum_partial(const void* start, uint32_t size);
+uint32_t pv_checksum_partial(const void* start, uint32_t size);
 uint16_t pv_checksum_finalise(uint32_t checksum);
 
 #ifdef __cplusplus
